+++ systemdata/projekter.php
@@ -1,32 +1,34 @@
 <?php
-// -----------------systemdata/projekter.php----lap 3.6.2---2016-01-18----
+//                ___   _   _   ___  _     ___  _ _
+//               / __| / \ | | |   \| |   |   \| / /
+//               \__ \/ ^ \| |_| |) | | _ | |) |  <
+//               |___/_/ \_|___|___/|_||_||___/|_\_\
 //
-// Dette program er fri software. Du kan gendistribuere det og / eller
-// modificere det under betingelserne i GNU General Public License (GPL)
-// som er udgivet af The Free Software Foundation; enten i version 2
-// af denne licens eller en senere version efter eget valg
-// Fra og med version 3.2.2 dog under iagttagelse af følgende:
-// 
-// Programmet må ikke uden forudgående skriftlig aftale anvendes
-// i konkurrence med DANOSOFT ApS eller anden rettighedshaver til programmet.
+// --- systemdata/projekter.php --- lap 3.9.9 --- 2021-02-11 ---
+// LICENSE
 //
-// Dette program er udgivet med haab om at det vil vaere til gavn,
-// men UDEN NOGEN FORM FOR REKLAMATIONSRET ELLER GARANTI. Se
-// GNU General Public Licensen for flere detaljer.
+// This program is free software. You can redistribute it and / or
+// modify it under the terms of the GNU General Public License (GPL)
+// which is published by The Free Software Foundation; either in version 2
+// of this license or later version of your choice.
+// However, respect the following:
 //
-// En dansk oversaettelse af licensen kan laeses her:
-// http://www.saldi.dk/dok/GNU_GPL_v2.html
+// It is forbidden to use this program in competition with Saldi.DK ApS
+// or other proprietor of the program without prior written agreement.
 //
-// Copyright (c) 2004-2016 DANOSOFT ApS
+// The program is published with the hope that it will be beneficial,
+// but WITHOUT ANY KIND OF CLAIM OR WARRANTY. See
+// GNU General Public License for more details.
 //
+// Copyright (c) 2003-2020 saldi.dk aps
 // ----------------------------------------------------------------------
 // 20160118 div smårettelser.
-//
+// 20210211 PHR Some cleanup
 
 @session_start();
 $s_id=session_id();
 
-$nopdat=NULL;	
+$cfg = $nopdat = NULL;	
 
 $modulnr=1;
 $title="Systemsetup";
@@ -75,10 +77,12 @@ if ($gem) {
 	db_modify("delete from grupper where art='PRJ'and id = $id",__FILE__ . " linje " . __LINE__);
 	$id=0;
 }
-$r=db_fetch_array(db_select("SELECT * FROM grupper where beskrivelse='projekt' and art = 'PRJ' and kodenr = '0' order by id",__FILE__ . " linje " . __LINE__));
+$qtxt="SELECT * FROM grupper where beskrivelse='projekt' and art = 'PRJ' and kodenr = '0' order by id";
+if ($r=db_fetch_array(db_select($qtxt,__FILE__ . " linje " . __LINE__))) {
 $cfg_id=$r['id']*1;
 $cfg=$r['box1'];
 db_modify("delete from grupper where art='PRJ' and kodenr='0' and id != $cfg_id",__FILE__ . " linje " . __LINE__);
+}
 if (!$cfg) $cfg=4;
 /*
 if (!$cfg) {
@@ -115,6 +119,8 @@ function tilpas($id,$cfg) {
 function rediger($id) {
 	global $cfg;
 	global $db_encode;
+	
+	$projektnr = $pos =
 	
 	$id*=1;
 	if ($id) {
